refactoring the base classes
