+++ modules/tool_call/common.py
@@ -2,6 +2,7 @@ import subprocess
 from ast import literal_eval
 import json
 
+from modules.common.base_tool import BaseTool as GenericTool
 
 class RunGraphMixin():
     def _run_graph(self, *args):
@@ -18,19 +19,4 @@ class RunGraphMixin():
         last_row = ret.rstrip().split("\n")[-1]
         v1 = json.loads(last_row)
         return v1
-
-class GenericTool():
-	def __init__(self):
-		pass
-
-	def _parse_inputs(self,fname, text_params):
-		if isinstance(text_params,list):
-			return text_params,{}
-		val = """def args_to_dict(*args,**kwargs):\n  return args,kwargs\nres = args_to_dict({})\n"""
-		ex = val.format(text_params)
-		l = {}
-		exec(ex, l, l)
-		a,k = l["res"]
-		b=list(a)
-		return b,k
